address code review and reorganize files
FU TS
