+++ packages/arcgis-rest-feature-service/src/index.ts
@@ -1,1 +1,4 @@
-export * from "./features";
+export * from "./query";
+export * from "./add";
+export * from "./update";
+export * from "./delete";
